@@ -1,8 +1,8 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 29
+    buildToolsVersion '29.0.2'
     flavorDimensions 'default'
     project.ext.set('gapps', true)
     compileOptions {
@@ -11,10 +11,10 @@ android {
     }
     defaultConfig {
         applicationId 'org.exobel.routerkeygen'
-        versionCode 73
-        versionName '4.1.7'
+        versionCode 80
+        versionName '4.2.0'
         minSdkVersion 16
-        targetSdkVersion 27
+        targetSdkVersion 29
         multiDexEnabled false
     }
     buildTypes {
@@ -40,7 +40,7 @@ android {
     // call regular ndk-build(.cmd) script from app directory
     task ndkBuild(type: Exec) {
         workingDir file('src/main')
-        commandLine getNdkBuildCmd()
+        commandLine getNdkBuildCmd(), 'V=1'
     }
     tasks.withType(JavaCompile) {
         compileTask -> compileTask.dependsOn ndkBuild
@@ -80,7 +80,7 @@ def getNdkDir() {
 }
 
 def getNdkBuildCmd() {
-    def ndkBuild = getNdkDir() + File.separator + 'ndk-build'
+    def ndkBuild = getNdkDir() + File.separator + 'build' + File.separator + 'ndk-build'
     if (org.gradle.internal.os.OperatingSystem.current().isWindows()) {
         ndkBuild += '.cmd'
     }
@@ -90,11 +90,11 @@ def getNdkBuildCmd() {
 
 
 dependencies {
-    compile project(':aFileChooser')
-    implementation 'com.android.support:support-compat:27.0.1'
-    implementation 'com.android.support:support-core-ui:27.0.1'
-    implementation 'com.android.support:support-core-utils:27.0.1'
-    implementation 'com.android.support:support-fragment:27.0.1'
+    implementation project(':aFileChooser')
+    implementation 'androidx.core:core:1.1.0'
+    implementation 'androidx.legacy:legacy-support-core-ui:1.0.0'
+    implementation 'androidx.legacy:legacy-support-core-utils:1.0.0'
+    implementation 'androidx.fragment:fragment:1.2.0'
 
     if (gapps) {
         implementation('com.google.android.gms:play-services-location:11.6.0', {
