@@ -18,7 +18,7 @@ package com.ipaulpro.afilechooser;
 
 import android.content.Context;
 import android.os.FileObserver;
-import android.support.v4.content.AsyncTaskLoader;
+import androidx.loader.content.AsyncTaskLoader;
 
 import com.ipaulpro.afilechooser.utils.FileUtils;
 
