@@ -4,9 +4,7 @@ import android.net.wifi.ScanResult;
 import android.net.wifi.WifiManager;
 import android.os.Parcel;
 import android.os.Parcelable;
-import android.support.annotation.NonNull;
-
-import com.farproc.wifi.connecter.Wifi;
+import androidx.annotation.NonNull;
 
 import org.exobel.routerkeygen.WirelessMatcher;
 
