@@ -22,7 +22,7 @@ package org.exobel.routerkeygen.ui;
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v4.app.NavUtils;
+import androidx.core.app.NavUtils;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.widget.Toast;
